+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.1.0 - 2021-05-04
+# Changelog
 
-* Extract into independent package
+## [0.1.1] - 2021-05-06
+
+* Add `child` to `Beacon`
+
+## [0.1.0] - 2021-05-04
+
+* Extract code into independent package